Move context processors into separate module to avoid confusion
In some places context processors were being called in the expectation
that they would return a boolean response.
