+++ mtp_noms_ops/apps/security/views/nomis.py
@@ -7,13 +7,13 @@ from django.utils.cache import patch_cache_control
 from mtp_common.nomis import get_photograph_data, get_location
 from requests.exceptions import RequestException
 
-from security.utils import nomis_api_available
+from security.utils import is_nomis_api_configured
 
 logger = logging.getLogger('mtp')
 
 
 def prisoner_image_view(request, prisoner_number):
-    if nomis_api_available(request) and prisoner_number:
+    if is_nomis_api_configured() and prisoner_number:
         try:
             b64data = get_photograph_data(prisoner_number)
             if b64data:
@@ -32,7 +32,7 @@ def prisoner_image_view(request, prisoner_number):
 
 def prisoner_nomis_info_view(request, prisoner_number):
     response_data = {}
-    if nomis_api_available(request) and prisoner_number:
+    if is_nomis_api_configured() and prisoner_number:
         try:
             location = get_location(prisoner_number)
             if 'housing_location' in location:
