+++ mtp_noms_ops/settings/base.py
@@ -125,11 +125,11 @@ TEMPLATES = [
                 'mtp_common.context_processors.app_environment',
                 'mtp_noms_ops.utils.govuk_localisation',
                 'mtp_noms_ops.utils.external_breadcrumbs',
-                'mtp_noms_ops.apps.security.utils.initial_params',
-                'mtp_noms_ops.apps.security.utils.initial_disbursement_params',
-                'mtp_noms_ops.apps.security.utils.nomis_api_available',
-                'mtp_noms_ops.apps.security.utils.prison_choice_available',
-                'mtp_noms_ops.apps.security.utils.notifications_available',
+                'mtp_noms_ops.apps.security.context_processors.initial_params',
+                'mtp_noms_ops.apps.security.context_processors.initial_disbursement_params',
+                'mtp_noms_ops.apps.security.context_processors.nomis_api_available',
+                'mtp_noms_ops.apps.security.context_processors.prison_choice_available',
+                'mtp_noms_ops.apps.security.context_processors.notifications_available',
             ],
         },
     },
